# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
    "/auth/register": {
      "post": {
        "operationId": "AuthController_register",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RegisterDto"
              },
              "examples": {
                "default": {
                  "summary": "New company + admin (no plan → trial)",
                  "value": {
                    "companyName": "Acme Operations",
                    "email": "admin@acme.com",
                    "password": "Passw0rd!",
                    "establishmentNumber": "1-2345678"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/login": {
      "post": {
        "operationId": "AuthController_login",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LoginDto"
              },
              "examples": {
                "default": {
                  "summary": "Login",
                  "value": {
                    "email": "admin@acme.com",
                    "password": "Passw0rd!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/refresh": {
      "post": {
        "operationId": "AuthController_refresh",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "cookie": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/logout": {
      "post": {
        "operationId": "AuthController_logout",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/change-password": {
      "post": {
        "operationId": "AuthController_changePassword",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ChangePasswordDto"
              },
              "examples": {
                "default": {
                  "summary": "Change password",
                  "value": {
                    "currentPassword": "Passw0rd!",
                    "newPassword": "NewPassw0rd!"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
    "/auth/me": {
      "get": {
        "operationId": "AuthController_me",
        "parameters": [],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Auth"
        ]
      }
    },
...
    "/attendance/check-in": {
      "post": {
        "operationId": "AttendanceController_checkIn",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckInDto"
              },
              "examples": {
                "owner": {
                  "summary": "Owner punches for an employee",
                  "value": {
                    "employeeId": "employee-uuid"
                  }
                },
                "portal": {
                  "summary": "Employee self check-in (empty body)",
                  "value": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/check-out": {
      "post": {
        "operationId": "AttendanceController_checkOut",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CheckOutDto"
              },
              "examples": {
                "owner": {
                  "summary": "Owner punches out for an employee",
                  "value": {
                    "employeeId": "employee-uuid"
                  }
                },
                "portal": {
                  "summary": "Employee self check-out",
                  "value": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance": {
      "post": {
        "operationId": "AttendanceController_upsert",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpsertAttendanceDto"
              },
              "examples": {
                "default": {
                  "summary": "Manual entry (needs shift on employee)",
                  "value": {
                    "employeeId": "employee-uuid",
                    "date": "2026-06-29",
                    "checkIn": "2026-06-29T08:05:00+03:00",
                    "checkOut": "2026-06-29T17:30:00+03:00"
                  }
                },
                "leave": {
                  "summary": "Mark a leave day",
                  "value": {
                    "employeeId": "employee-uuid",
                    "date": "2026-06-29",
                    "status": "LEAVE"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      },
      "get": {
        "operationId": "AttendanceController_findAll",
        "parameters": [
          {
            "name": "employeeId",
            "required": false,
            "in": "query",
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PRESENT",
                "ABSENT",
                "LEAVE"
              ]
            }
          },
          {
            "name": "dateFrom",
            "required": false,
            "in": "query",
            "description": "Inclusive lower bound (date).",
            "schema": {
              "example": "2026-06-01",
              "type": "string"
            }
          },
          {
            "name": "dateTo",
            "required": false,
            "in": "query",
            "description": "Inclusive upper bound (date).",
            "schema": {
              "example": "2026-06-30",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/bulk": {
      "post": {
        "operationId": "AttendanceController_bulk",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/BulkAttendanceDto"
              },
              "examples": {
                "default": {
                  "summary": "Two rows",
                  "value": {
                    "records": [
                      {
                        "employeeId": "employee-uuid",
                        "date": "2026-06-29",
                        "checkIn": "2026-06-29T08:00:00+03:00",
                        "checkOut": "2026-06-29T17:10:00+03:00"
                      },
                      {
                        "employeeId": "employee-uuid-2",
                        "date": "2026-06-29",
                        "status": "ABSENT"
                      }
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
    "/attendance/{id}": {
      "get": {
        "operationId": "AttendanceController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      },
      "patch": {
        "operationId": "AttendanceController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateAttendanceDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "tags": [
          "Attendance"
        ]
      }
    },
...
  },
  "info": {
    "title": "HR System API",
    "description": "Multi-tenant Payroll & HR Management API (Phases 1–6). Import into APIDog via live URL http://localhost:3004/docs-json or backend/postman_collection.json. Swagger UI: http://localhost:3004/docs",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [
    {
      "url": "http://localhost:3004/api",
      "description": "Local"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearer": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http"
      },
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "refreshToken"
      }
    },
    "schemas": {
      "RegisterDto": {
        "type": "object",
        "properties": {
          "companyName": {
            "type": "string",
            "example": "Acme Operations",
            "maxLength": 120
          },
          "email": {
            "type": "string",
            "example": "admin@acme.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!",
            "minLength": 8,
            "maxLength": 72,
            "description": "Must contain at least one lowercase letter, one uppercase letter and one number"
          },
          "establishmentNumber": {
            "type": "string",
            "example": "1-2345678",
            "maxLength": 50,
            "description": "MOL/WPS establishment number"
          },
          "planId": {
            "type": "string",
            "format": "uuid",
            "description": "Optional. Omit to create the company with NO plan (stays on TRIAL)."
          }
        },
        "required": [
          "companyName",
          "email",
          "password"
        ]
      },
      "LoginDto": {
        "type": "object",
        "properties": {
          "email": {
            "type": "string",
            "example": "admin@acme.com",
            "format": "email"
          },
          "password": {
            "type": "string",
            "example": "Passw0rd!"
          }
        },
        "required": [
          "email",
          "password"
        ]
      },
      "ChangePasswordDto": {
        "type": "object",
        "properties": {
          "currentPassword": {
            "type": "string",
            "example": "Passw0rd!"
          },
          "newPassword": {
            "type": "string",
            "example": "NewPassw0rd!",
            "minLength": 8,
            "maxLength": 72,
            "description": "Must contain at least one lowercase letter, one uppercase letter and one number"
          }
        },
        "required": [
          "currentPassword",
          "newPassword"
        ]
      },
...
      "CheckInDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "description": "Required for Company Owner. Portal employees must omit this (self only)."
          },
          "at": {
            "type": "string",
            "example": "2026-06-29T08:05:00+03:00",
            "description": "ISO timestamp (include offset or Z). Defaults to now."
          },
          "shiftId": {
            "type": "string",
            "format": "uuid",
            "description": "Override the employee's default shift for this day."
          }
        }
      },
      "CheckOutDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid",
            "description": "Required for Company Owner. Portal employees must omit this (self only)."
          },
          "at": {
            "type": "string",
            "example": "2026-06-29T17:30:00+03:00",
            "description": "ISO timestamp (include offset or Z). Defaults to now."
          }
        }
      },
      "UpsertAttendanceDto": {
        "type": "object",
        "properties": {
          "employeeId": {
            "type": "string",
            "format": "uuid"
          },
          "date": {
            "type": "string",
            "example": "2026-06-29",
            "description": "Calendar day (tenant-local)."
          },
          "shiftId": {
            "type": "string",
            "format": "uuid"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-06-29T08:05:00+03:00"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-06-29T17:30:00+03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "ABSENT",
              "LEAVE"
            ],
            "example": "PRESENT"
          },
          "delayMinutes": {
            "type": "number",
            "minimum": 0,
            "description": "Manual override of derived delay."
          },
          "overtimeHours": {
            "type": "number",
            "minimum": 0,
            "description": "Manual override of derived overtime."
          }
        },
        "required": [
          "employeeId",
          "date"
        ]
      },
      "BulkAttendanceDto": {
        "type": "object",
        "properties": {
          "records": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/UpsertAttendanceDto"
            }
          }
        },
        "required": [
          "records"
        ]
      },
      "UpdateAttendanceDto": {
        "type": "object",
        "properties": {
          "shiftId": {
            "type": "string",
            "format": "uuid"
          },
          "checkIn": {
            "type": "string",
            "example": "2026-06-29T08:05:00+03:00"
          },
          "checkOut": {
            "type": "string",
            "example": "2026-06-29T17:30:00+03:00"
          },
          "status": {
            "type": "string",
            "enum": [
              "PRESENT",
              "ABSENT",
              "LEAVE"
            ]
          },
          "delayMinutes": {
            "type": "number",
            "minimum": 0
          },
          "overtimeHours": {
            "type": "number",
            "minimum": 0
          }
        }
      },
...
    }
  }
}